[lc_ctrl,dv] Tighten type of read_lc_state
This was already returning a value <= 0xff. Reflect that in the type.

+++ src/integration/test_suites/libs/lc_ctrl/lc_ctrl.c
@@ -363,7 +363,7 @@ void force_PPD_pin(void) {
     VPRINTF(LOW, "MCU: RMA_SCRAP_PPD pin asserted high!\n");
 }
 
-uint32_t read_lc_state(void) {
+uint8_t read_lc_state(void) {
     for (uint32_t i = 0; i < 512; i++) {
         __asm__ volatile ("nop");
     }
